Fix issue with 1xn dummy count matrix not working with CreateSeuratObject
With Seurat 5.0 at least, a single-row count matrix does not work for CreateSeuratObject(). Also simplified the creation of the matrix to avoid having to create it and then transpose. The matrix is changed to a sparse matrix by Seurat, so adding some rows doesn't affect the memory profile much.
Tested with Seurat 5.0 in R 4.3.2